+++ Source/Core/Transforms.js
@@ -338,12 +338,8 @@ define([
             }
 
             var tmp = pointToWindowCoordinatesTemp;
-            tmp.x = point.x;
-            tmp.y = point.y;
-            tmp.z = point.z;
-            tmp.w = 1.0;
 
-            Matrix4.multiplyByVector(modelViewProjectionMatrix, tmp, tmp);
+            Matrix4.multiplyByPoint(modelViewProjectionMatrix, point, tmp);
             Cartesian4.multiplyByScalar(tmp, 1.0 / tmp.w, tmp);
             Matrix4.multiplyByVector(viewportTransformation, tmp, tmp);
             return Cartesian2.fromCartesian4(tmp, result);